// frontend/src/main/frontend/src/app/model/CompetitionList.ts
export class CompetitionList {
  hasBaroque: boolean = false;
  baroqueStart: Date;
  baroqueEnd: Date;
  hasJazz: boolean = false;
  jazzStart: Date;
  jazzEnd: Date;
  hasFree: boolean = false;
  freeStart: Date;
  freeEnd: Date;
  hasComposition:boolean = false;
  compositionStart: Date;
  compositionEnd: Date;
  hasConcert:boolean = false;
  concertStart: Date;
  concertEnd: Date;
}